requirements.txt
Appium-Python-Client==5.2.2
attrs==25.3.0
behave==1.3.1
certifi==2025.8.3
cffi==1.17.1
colorama==0.4.6
cucumber-expressions==18.0.1
cucumber-tag-expressions==6.2.0
h11==0.16.0
idna==3.10
outcome==1.3.0.post0
parse==1.20.2
parse_type==0.6.6
psutil==7.0.0
pycparser==2.22
PySocks==1.7.1
selenium==4.35.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
urllib3==2.5.0
websocket-client==1.8.0
wsproto==1.2.0
